Ref: extracting and reusing checks.
